requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
flask-marshmallow==1.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.3.0
marshmallow-sqlalchemy==1.5.0
mysql-connector-python==9.6.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
flask-limiter==3.12
flask-caching==2.3.1
python-jose==3.5.0
